+++ src/main/java/com/dao/UserDAO.java
@@ -164,7 +164,7 @@ public class UserDAO {
     private User createUserEntity(ResultSet rs) throws SQLException {
 
         return new User(
-                rs.getInt("id"),
+                rs.getLong("id"),
                 rs.getString("user_name"),
                 rs.getBoolean("is_admin"));
     }
